Alt Version of collision class, only currently working for IOS probably. Some changes listed below
changed CalculateNearestNode to return Vector2 instead of assigning class variable.
implemented equals in vector 2
implemented isLinked in graph
In graph.findclosestNode, i checked for node validity in the graph list of vertices
plus a bunch more

This branch can be useful for just comparisons and shit in the final report.